Refactored LoggerAction to shared map.
Refactored LoggerAdapter into Service
