+++ src/main/java/demo/codeexample/logger/application/LoggerService.java
@@ -1,12 +1,10 @@
 package demo.codeexample.logger.application;
 
 import demo.codeexample.logger.LoggerLookup;
-import demo.codeexample.logger.domain.LoggerAction;
+import demo.codeexample.shared.LoggerAction;
 import demo.codeexample.logger.domain.LoggerEntity;
 import demo.codeexample.logger.domain.LoggerRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -15,6 +13,7 @@ public class LoggerService implements LoggerLookup {
 
     private final LoggerRepository loggerRepository;
 
+    @Override
     public void log(LoggerAction action, Long userId, String entityType, Long entityId) {
         LoggerEntity log = new LoggerEntity();
         log.setAction(action);
