+++ src/main/java/demo/codeexample/shared/LoggerAction.java
@@ -1,4 +1,4 @@
-package demo.codeexample.logger.domain;
+package demo.codeexample.shared;
 
 public enum LoggerAction {
     PROJECT_CREATED,
